Extract statement/term/suggestion helpers methods
These will be useful to included in other comparisons classes too.
